Nuevas mejoras

+++ src/main/java/com/mayab/desarrollo/persistencia/DAOUsuario.java
@@ -12,6 +12,8 @@ public class DAOUsuario implements IDAOUsuario{
     public int insertarUsuario(Usuario usuario) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
+        int hueco = this.obtenerHueco();
+        session.createNativeQuery(String.format("UPDATE hibernate_sequence SET next_val = %d" , hueco)).executeUpdate();
         int id = (int)session.save(usuario);
         session.getTransaction().commit();
 		session.close();
@@ -57,11 +59,23 @@ public class DAOUsuario implements IDAOUsuario{
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
         Usuario usuarioBorrar = this.consultarUsuario(id);
-        session.delete(usuarioBorrar);
-        session.getTransaction().commit();
-        session.close();
-        return this.consultarUsuario(id) == null;
+        if(usuarioBorrar != null){
+            session.delete(usuarioBorrar);
+            session.getTransaction().commit(); 
+            if(this.listarTodosLosUsuarios().isEmpty()){
+                session.beginTransaction(); 
+                session.createNativeQuery("UPDATE hibernate_sequence SET next_val = 1").executeUpdate();
+                System.out.println("Se quedó vacia");
+                session.getTransaction().commit();
+            }
+            session.close();
+            return true;
+        } else {
+            session.close();
+            return false;
+        }
     }
+    
 
   
 
@@ -70,7 +84,7 @@ public class DAOUsuario implements IDAOUsuario{
        
         Session session = HibernateUtil.getSessionFactory().openSession();
         session.beginTransaction();
-        List<Usuario> todosUsuarios = session.createQuery("FROM Usuario", Usuario.class).getResultList();
+        List<Usuario> todosUsuarios = session.createNativeQuery("SELECT * FROM Usuarios ORDER BY id", Usuario.class).getResultList();
         session.getTransaction().commit();
         session.close();
         return todosUsuarios;
@@ -79,7 +93,6 @@ public class DAOUsuario implements IDAOUsuario{
     @Override
     public Usuario validarUsuario(String usuario, String email) {
         List<Usuario> usuarios = listarTodosLosUsuarios();
-    
         for (Usuario user : usuarios) {
             if(user.getNombre().equals(usuario) || user.getEmail().equals(email) ){
                 return user;
@@ -87,6 +100,23 @@ public class DAOUsuario implements IDAOUsuario{
         }
 
         return null;
+    }
+
+    @Override
+    public int obtenerHueco() {
+        int hueco = 0;
+        List<Usuario> usuarios = listarTodosLosUsuarios();
+        int maxId = usuarios.getLast().getId();
+        if(usuarios.size() != maxId){
+            for (Usuario user : usuarios) {
+                hueco++;
+                if(user.getId() != hueco){
+                    break;
+                }
+            }
+            return hueco;
+        }
+        return (maxId + 1);
     }
     
 }
